song tagging database functions
Adds database functions to simplify song tagging upon upload. The base function, add_song_tag, updates the tables for a single added song-tag relationship. The other function, add_many_song_tags, uses the base function to add a bunch of tags at once for a single song (as long as those tags are in the same category). That should reduce the number of network calls and abstract away some of the complexity from the client-side app a bit. :)